🚚 Rename validation.Validator to Validator.from_df
