--- lamindb/validation/__init__.py
@@ -1,19 +1,0 @@
-"""Validators built on LaminDB.
-
-Import the package::
-
-   from lamindb.validation import Validator, AnnDataValidator
-
-This is the complete API reference:
-
-.. autosummary::
-   :toctree: .
-
-   Validator
-   AnnDataValidator
-   Lookup
-"""
-
-from ._anndata_validator import AnnDataValidator
-from ._lookup import Lookup
-from ._validator import Validator
